Test 2 only allow post from kaha's domain
